use binary input feature

+++ libdeepgalois/include/deepgalois/types.h
@@ -18,7 +18,7 @@ typedef std::vector<vec_t>
 typedef std::vector<feature_t> FV; // feature vector
 typedef std::vector<FV> FV2D;      // feature vectors: num_samples x feature_dim
 typedef float acc_t;               // Accuracy type
-typedef short label_t;  // label is for classification (supervised learning)
+typedef uint8_t label_t;  // label is for classification (supervised learning)
 typedef uint8_t mask_t; // mask is used to indicate different uses of labels:
                         // train, val, test
 typedef uint32_t VertexID;
